+++ backend/src/main/java/de/learnlib/alex/modelchecking/entities/LtsCheckingResult.java
@@ -19,7 +19,6 @@ package de.learnlib.alex.modelchecking.entities;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import net.automatalib.words.Word;
 
-import java.util.ArrayList;
 import java.util.List;
 
 /** The counterexample of a model check. */
@@ -28,6 +27,12 @@ public class LtsCheckingResult {
     /** The forumla that has been checked. */
     private LtsFormula formula;
 
+    /** The ID of the learner result. */
+    private Long learnerResultId;
+
+    /** Which hypothesis has been checked. */
+    private Long learnerResultStepNo;
+
     /** The prefix of the counterexample. */
     private List<String> prefix;
 
@@ -35,22 +40,14 @@ public class LtsCheckingResult {
     private List<String> loop;
 
     /** Constructor. */
-    public LtsCheckingResult(LtsFormula formula) {
-        this.formula = formula;
-        this.prefix = new ArrayList<>();
-        this.loop = new ArrayList<>();
+    public LtsCheckingResult(LtsFormula formula, Long learnerResultId, Long learnerResultStepNo) {
+        this(formula, learnerResultId, learnerResultStepNo, Word.epsilon(), Word.epsilon());
     }
 
-    /**
-     * Constructor.
-     *
-     * @param prefix
-     *         {@link #prefix}
-     * @param loop
-     *         {@link #loop}
-     */
-    public LtsCheckingResult(LtsFormula formula, Word<String> prefix, Word<String> loop) {
+    public LtsCheckingResult(LtsFormula formula, Long learnerResultId, Long learnerResultStepNo, Word<String> prefix, Word<String> loop) {
         this.formula = formula;
+        this.learnerResultId = learnerResultId;
+        this.learnerResultStepNo = learnerResultStepNo;
         this.prefix = prefix.asList();
         this.loop = loop.asList();
     }
@@ -77,6 +74,22 @@ public class LtsCheckingResult {
 
     public void setFormula(LtsFormula formula) {
         this.formula = formula;
+    }
+
+    public Long getLearnerResultId() {
+        return learnerResultId;
+    }
+
+    public void setLearnerResultId(Long learnerResultId) {
+        this.learnerResultId = learnerResultId;
+    }
+
+    public Long getLearnerResultStepNo() {
+        return learnerResultStepNo;
+    }
+
+    public void setLearnerResultStepNo(Long learnerResultStepNo) {
+        this.learnerResultStepNo = learnerResultStepNo;
     }
 
     /**